requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
click==8.1.3
cmake==3.24.1.1
Django==4.1.2
djangorestframework==3.14.0
face-recognition-models==0.3.0
Pillow==9.2.0
pytz==2022.4
sqlparse==0.4.3
dlib
